att 03 mudança nas classes no quesito id, de string para long

+++ esg/esg/src/main/java/br/com/fiap/esg/model/ParticipacaoTreinamento.java
@@ -14,7 +14,7 @@ public class ParticipacaoTreinamento {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEQ_PARTICIPACAO")
     @SequenceGenerator(name = "SEQ_PARTICIPACAO", sequenceName = "SEQ_PARTICIPACAO", allocationSize = 1)
     @Column(name = "idParticipacao")
-    private Integer idParticipacao;
+    private Long idParticipacao;
     private LocalDate dataParticipacao;
     private String status;
 }
